requirements.txt
lightning==2.1.3
numpy==1.26.2
opencv-python==4.8.1.78
pillow==9.4.0
pytorch-lightning==2.1.3
pyyaml==6.0.1
torch==2.1.2
torchaudio==2.1.2
torchmetrics==1.2.1
torchvision==0.16.2
tqdm==4.66.1
wandb==0.16.1
yacs==0.1.8
